refactor(ai-sdk-params): extract mobile agent parameter assembly
